fix(nextjs): Should be able to run custom server targets with swc (#27526)
## Currently
When we installed swc from the Next.js generator it installed an older
version that expected version of swc this caused the node & build
generator to not work.

## Expected
With these changes the version will be in sync with what the
`@nx/js:swc` executor is expecting. With that in mind the build should
now work as expected and by extension the serve target for custom server
as well.

fixes: #27222
(cherry picked from commit 554e979abfaef3045f98d5bae55843df87776a4d)