fix(react): tailwind with rspack should work on generation #27028 (#27619)
<!-- Please make sure you have read the submission guidelines before
posting an PR -->
<!--
https://github.com/nrwl/nx/blob/master/CONTRIBUTING.md#-submitting-a-pr
-->

<!-- Please make sure that your commit message follows our format -->
<!-- Example: `fix(nx): must begin with lowercase` -->

<!-- If this is a particularly complex change or feature addition, you
can request a dedicated Nx release for this pull request branch. Mention
someone from the Nx team or the `@nrwl/nx-pipelines-reviewers` and they
will confirm if the PR warrants its own release for testing purposes,
and generate it for you if appropriate. -->

## Current Behavior
<!-- This is the behavior we have today -->
React, Tailwind and Rspack is not working out of the box

## Expected Behavior
<!-- This is the behavior we should expect with the changes in this PR
-->
It works out of the box.

In tandem with nrwl/nx-labs#415

## Related Issue(s)
<!-- Please link the issue being fixed so it gets closed when this is
merged. -->

Fixes #27028

(cherry picked from commit add4183d0a9acebd4ba7e28674a4597281c55828)